[core] Move runWithLock to jasmine_util
